requirements.txt
altair
annotated-types
anyio
anytree
APScheduler
area
asttokens
attrs
bcrypt
beautifulsoup4
bleach
blinker
bokeh
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
colorama
contourpy
cryptography
cycler
dash
dash-core-components
dash-cytoscape
dash-html-components
dash-table
dask
dask-expr
decorator
duckdb
entrypoints
executing
Faker
fastapi
favicon
Flask
fonttools
fsspec
gitdb
GitPython
greenlet
htbuilder
idna
importlib_metadata
ipython
itsdangerous
jedi
Jinja2
jsonpickle
jsonschema
jsonschema-specifications
kiwisolver
linkify-it-py
locket
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
narwhals
neo4j
nest-asyncio
networkx
numpy
oracledb
packaging
pandas
panel
param
parso
partd
pillow
plotly
prettytable
progressbar2
prometheus_client
prompt_toolkit
protobuf
psutil
psycopg2
pure_eval
pyarrow
pycoral
pycparser
pydantic
pydantic_core
pydeck
pyecharts
Pygments
pymdown-extensions
pyparsing
PyQt5
PyQt5-Qt5
PyQt5_sip
pyqtgraph
python-dateutil
python-decouple
python-utils
pytz
pyvis
pyviz_comms
PyYAML
referencing
requests
retrying
rich
rpds-py
schedule
simplejson
six
smmap
sniffio
soupsieve
SQLAlchemy
st-annotated-text
st-theme
stack-data
starlette
streamlit
streamlit-ace
streamlit-aggrid
streamlit-autorefresh
streamlit-camera-input-live
streamlit-card
streamlit-echarts
streamlit-elements
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-option-menu
streamlit-plotly-events
streamlit-sortables
streamlit-toggle-switch
streamlit-tree-select
streamlit-vertical-slider
tenacity
toml
toolz
tornado
tqdm
traitlets
typing_extensions
tzdata
tzlocal
uc-micro-py
urllib3
validators
watchdog
wcwidth
webencodings
Werkzeug
xyzservices
zipp
oracledb